backend/requirements.txt
setuptools==68.0.0
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
mysqlclient==2.2.4
eth-account==0.11.0
eth-abi==6.0.0b1
eth-keyfile==0.9.1
eth-keys==0.7.0
eth-rlp==1.0.1
eth-typing==5.2.1
eth-utils==5.3.1
eth-hash[pycryptodome]==0.7.1
hexbytes==0.3.1
rlp==4.1.0
parsimonious==0.10.0
requests==2.31.0
python-dotenv==1.0.1
Pillow==10.2.0
reportlab==4.1.0
aiohttp==3.13.3
websockets==16.0
protobuf==7.34.0
pyunormalize==17.0.0
pywin32==311
jsonschema==4.26.0
pycryptodome==3.23.0
bitarray==3.8.0